refactor: delete previous Api file and remove unnecessary code
